ajout/suppr model. Changement nommage. Ajout de la vue + controller pour userSettings

+++ database/migrations/2022_09_27_080637_create_urssaf_settings_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('urssaf_settings', function (Blueprint $table) {
             $table->id();
-            $table->integer('percentage');
+            $table->float('percentage', 3, 2);
             $table->text('description');
             $table->timestamps();
         });
